ai-tutor-pairing/streamlit_apps/requirements.txt
streamlit==1.31.1
numpy==1.24.3
Pillow==10.0.0 